fixed a bug when deleting order with no transaction failed. now we check for transaction existance first.

+++ app/Order.php
@@ -47,19 +47,22 @@ class Order extends Model
     }
 
 
-
-    public static function boot ()
+    public static function boot()
     {
         parent::boot();
 
         self::deleting(function (Order $order) {
 
-            $order->transaction->delete();
+            if ($order->transaction) {
+                $order->transaction->delete();
+            }
 
         });
         self::restoring(function (Order $order) {
 
-            $order->transaction()->withTrashed()->first()->restore();
+            if ($order->transaction) {
+                $order->transaction()->withTrashed()->first()->restore();
+            }
 
         });
     }
